# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Caseworker Task Manager API
  description: REST API for creating, listing, updating, and deleting caseworker tasks.
  version: 1.2.0

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /api/health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/live:
    get:
      summary: Liveness probe
      description: Returns 200 if the process is running. No database check.
      operationId: getLive
      responses:
        "200":
          description: Process is alive

  /api/ready:
    get:
      summary: Readiness probe
      description: Returns 200 if the app can serve traffic (e.g. database reachable).
      operationId: getReady
      responses:
        "200":
          description: Ready to serve traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
        "503":
          description: Database or dependency unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/tasks:
    get:
      summary: List all tasks
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Number of tasks to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of tasks to skip before returning results.
      responses:
        "200":
          description: List of tasks ordered by due date
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"
    post:
      summary: Create a task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error (e.g. title required, dueAt in future)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Task UUID
    get:
      summary: Get a task by ID
      operationId: getTask
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid task ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"
    patch:
      summary: Update task status
      operationId: updateTaskStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error (e.g. invalid status, no change)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"
    put:
      summary: Update task fields
      operationId: updateTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"
    delete:
      summary: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    ReadyResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ready

    Task:
      type: object
      required: [id, title, status, dueAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        owner:
          type: string
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: string
          format: date-time
          description: RFC3339
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    CreateTaskRequest:
      type: object
      required: [title, status, dueAt]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        owner:
          type: string
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: string
          format: date-time
          description: RFC3339; must be in the future

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        tags:
          type: array
          items:
            type: string

    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [todo, in_progress, done]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message

  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthNotConfigured:
      description: API auth is required but API_AUTH_TOKEN is not configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque